function floor which takes a float n as argument and returns the floor of the float.

+++ 0x00-python_variable_annotations/2-floor.py
@@ -4,7 +4,8 @@
 """
 
 
-def floor(n: float) -> float:
+
+def floor(n: float) -> int:
     """Get the value of the floor of the float.
 
     Args:
